+++ projects/f-core/src/mixins/change-zoom/i-can-change-zoom.ts
@@ -1,0 +1,10 @@
+import { IPoint } from '../../types';
+
+export interface ICanChangeZoom {
+
+  setZoom(value: number, toPosition: IPoint): void;
+
+  setScalePosition(value: IPoint): void;
+
+  resetZoom(): void;
+}