+++ projects/f-core/src/mixins/fit-to-parent/i-can-fit-to-parent.ts
@@ -1,0 +1,7 @@
+import { IPoint, IRect } from '../../types';
+
+export interface ICanFitToParent {
+
+  fitToParent(rect: IRect, parentRect: IRect, points: IPoint[]): void;
+}
+