+++ projects/f-core/src/mixins/i-transformable.ts
@@ -1,0 +1,6 @@
+import { ITransformModel } from '../types';
+
+export interface ITransformable {
+
+  transform: ITransformModel;
+}