+++ projects/f-core/src/mixins/index.ts
@@ -1,0 +1,13 @@
+export * from './change-zoom';
+
+export * from './fit-to-parent';
+
+export * from './one-to-one-centering';
+
+export * from './change-position';
+
+export * from './constructor';
+
+export * from './i-transformable';
+
+