+++ projects/f-core/src/mixins/one-to-one-centering/i-can-one-to-one-centering.ts
@@ -1,0 +1,6 @@
+import { IPoint, IRect } from '../../types';
+
+export interface ICanOneToOneCentering {
+
+  oneToOneCentering(rect: IRect, parentRect: IRect, points: IPoint[]): void;
+}