+++ projects/f-core/src/public-api.ts
@@ -6,5 +6,7 @@ export * from './platform';
 
 export * from './infrastructure';
 
+export * from './mixins';
+
 export * from './types';