+++ projects/f-core/CHANGELOG.md
@@ -1,4 +1,16 @@
-## [1.0.8] - 2023-11-20
+## [1.0.8] - 2024-03-12
 ### Added
+
+- `changeZoom mixin` This mixin provides a way to change the zoom level of an element. 
+
+- `fitToParent mixin` This mixin provides a way to fit an element to its parent container.
+
+- `oneToOne mixin` This mixin provides a way to set the zoom level of an element to 1:1.
+
+- `changePosition mixin` This mixin provides a way to change the position of an element.
+
+## [1.0.7] - 2023-11-20
+### Added
+
 - `sanitizeElementId(id: string): string` function in the DOM Element module. This function provides cleaning and safe handling of DOM element IDs, preventing potential security and compatibility issues.